move geturls
